Autonomous and Teleop Differeniation Fixed
We fixed the error when we couldn't work on getting autonomous and
teleop to work. The error was that teleop worked perfectly fine but when
we activated autonomous, it sputtered A LOT and there was this Brownout
Protection error. What we did was change the battery which fixed the
brownout issue and then the same kind of slow movement in sputtering
occurred and we realized that the default command for tankdrive was
ALWAYS Running so it went against the auto and it changed the drive
system to 0 and then 1 and then 0, etc..... but we fixed it by adding a
requirement thing. yay